% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_param_variables.R
\name{assign_param_variables}
\alias{assign_param_variables}
\title{Assign PARAM_RAW values from input file to variable names}
\usage{
assign_param_variables(D, path)
}
\arguments{
\item{D}{xxxPARAMxxx}

\item{path}{xxxPARAMxxx}
}
\value{
PARAM list of all inputs from Excel template
}
\description{
initialize lists to hold all data/output
}
\details{
Template version

switches to use and filenames for TDL and Licor files

Prefix for output filenames

TIME WINDOW SECTION

OUTPUT SECTION

plot formats

interpolation:  1 = smooth_spline() like tanks
2 = use last reference mean until next measurement (horizontal interp) - for abrupt ref gas changes

Bootstrap

TDL sampling information

CONSTANTS SECTION

True calibration tank values

Which values to use

constants in formulas not given names

Calculation information

create a list to all returned parameters
}
