\name{calc.mean.Licor}
\alias{calc.mean.Licor}
\title{Calculate the mean for the Licor data based on last measurements}

\usage{calc.mean.Licor(TDL, Licor, TDL.cycle, sw)}
\arguments{
  \item{TDL}{}
  \item{Licor}{}
  \item{TDL.cycle}{}
  \item{sw}{}
}
\details{create Licor$summary for numerical summaries

Calculate mean for each site for retained observations}
\value{Licor}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





