\name{calc.mean.TDL.interp.tanks.ref}
\alias{calc.mean.TDL.interp.tanks.ref}
\title{Summary values for TDL interpolated tanks and reference}

\usage{calc.mean.TDL.interp.tanks.ref(TDL, TDL.cycle)}
\arguments{
  \item{TDL}{}
  \item{TDL.cycle}{}
}
\details{Calculate mean, var, and sd for each site for interpolated ranges.

Data updated to exclude indices that we don't summarize.}
\value{TDL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





