\name{calc.mean.TDL}
\alias{calc.mean.TDL}
\title{Calculate the mean and variance for the TDL/Licor data based on last measurements}

\usage{calc.mean.TDL(TDL, TDL.cycle)}
\arguments{
  \item{TDL}{}
  \item{TDL.cycle}{}
}
\details{create TDL$summary for numerical summaries

Calculate mean, var, and sd for each site for retained observations}
\value{TDL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





