% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_TDL_interp_tanks_ref.R
\name{calc_mean_TDL_interp_tanks_ref}
\alias{calc_mean_TDL_interp_tanks_ref}
\title{Title}
\usage{
calc_mean_TDL_interp_tanks_ref(TDL, TDL_cycle)
}
\arguments{
\item{TDL_cycle}{}
}
\description{
Title
}
