\name{extract.cycle.timing}
\alias{extract.cycle.timing}
\title{Cycle timing}

\usage{extract.cycle.timing(TDL, TDL.cycle, TDL.site.timing.filename)}
\arguments{
  \item{TDL}{}
  \item{TDL.cycle}{}
  \item{TDL.site.timing.filename}{}
}

\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





