\name{f.init.bs.matrix}
\alias{f.init.bs.matrix}
\title{init bs values to zero}

\usage{f.init.bs.matrix(n, R.bootstrap)}
\arguments{
  \item{n}{}
  \item{R.bootstrap}{}
}
\details{Create bs for calculated bootstrap values to estimate error, initializing to zeros.}
\value{bs}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





