\name{f.np.bs.mean}
\alias{f.np.bs.mean}
\title{calculate NP BS mean}

\usage{f.np.bs.mean(x, first.ind, last.ind, n.sam, n)}
\arguments{
  \item{x}{}
  \item{first.ind}{}
  \item{last.ind}{}
  \item{n.sam}{}
  \item{n}{}
}
\details{Resample from data with replacement, calculating mean of n nonparametric bootstrap resamples.}
\value{x.mean}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





