\name{f.par.bs.mean}
\alias{f.par.bs.mean}
\title{calculate Par BS mean}

\usage{f.par.bs.mean(x, x.var, n.sam, n)}
\arguments{
  \item{x}{}
  \item{x.var}{}
  \item{n.sam}{}
  \item{n}{}
}
\details{Sample from normal distribution with mean and se from the data, calculating mean of n parametric bootstrap resamples.}
\value{x.mean}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





