\name{f.plot.CI.individuals.driver}
\alias{f.plot.CI.individuals.driver}
\title{plot each individual variable}

\usage{f.plot.CI.individuals.driver(val, R.bootstrap, plot.format.list, 
    output.fn.prefix)}
\arguments{
  \item{val}{}
  \item{R.bootstrap}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{For every variable, creates plots using \code{\link{f.plot.CI.individuals}}.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





