\name{f.val.bs.CI.endpoints}
\alias{f.val.bs.CI.endpoints}
\title{create CI endpoints}

\usage{f.val.bs.CI.endpoints(x, n, ind.CI)}
\arguments{
  \item{x}{}
  \item{n}{}
  \item{ind.CI}{}
}
\details{Sorts BS samples, takes quantiles for equal-tailed sig.CI Level CIs.}
\value{x.CI}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





