\name{f.val.bs.matrix.check}
\alias{f.val.bs.matrix.check}
\title{Helper for \code{\link{f.val.bs.matrix}}, if the vector has zero length, return NA (for when using only TDL and not Licor)}

\usage{f.val.bs.matrix.check(val)}
\arguments{
  \item{val}{}
}

\value{val}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





