\name{f.val.bs.matrix}
\alias{f.val.bs.matrix}
\title{update values for each bs iterate}

\usage{f.val.bs.matrix(bs, val.calc.bs.temp, i.bs)}
\arguments{
  \item{bs}{}
  \item{val.calc.bs.temp}{}
  \item{i.bs}{}
}
\details{Move current value of \code{val.calc.bs.temp} into \code{bs$*} after checked with \code{\link{f.val.bs.matrix.check}}.}
\value{bs}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





