\name{f.val.calc.delta.13C.Resp}
\alias{f.val.calc.delta.13C.Resp}
\title{delta13C Respired, isotopic composition of respired CO2}

\usage{f.val.calc.delta.13C.Resp(reference.delta.e, chamber.delta.o, 
    p)}
\arguments{
  \item{reference.delta.e}{}
  \item{chamber.delta.o}{}
  \item{p}{}
}
\details{\deqn{delta.13C.Resp = (chamber.delta.o - reference.delta.e * (1 - p)) / p}}
\value{delta.13C.Resp}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





