\name{f.val.calc.Delta.i.simple.for.gm}
\alias{f.val.calc.Delta.i.simple.for.gm}
\title{Delta.i simple for gm, predicted discrimination including boundary layer effects but not decarboxylation effects}

\usage{f.val.calc.Delta.i.simple.for.gm(a, b, a.b, pa, ps, pi)}
\arguments{
  \item{a}{}
  \item{b}{}
  \item{a.b}{}
  \item{pa}{}
  \item{ps}{}
  \item{pi}{}
}
\details{\deqn{Delta.i.simple.for.gm = a.b * ((pa - ps) / pa) + a*((ps - pi) / pa) + b * (pi / pa)}}
\value{Delta.i.simple.for.gm}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





