\name{f.val.calc.Delta.i.simple.for.modeling}
\alias{f.val.calc.Delta.i.simple.for.modeling}
\title{Delta.i simple for modeling  a + (b-a) pi/pa predicted discrimination including boundary layer effects but using b adjustments to approximate effects of gm and decarboxylations}

\usage{f.val.calc.Delta.i.simple.for.modeling(a, b, pa, pi)}
\arguments{
  \item{a}{}
  \item{b}{}
  \item{pa}{}
  \item{pi}{}
}
\details{\deqn{Delta.i.simple.for.modeling = a + (b - a) * pi / pa}}
\value{Delta.i.simple.for.modeling}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





