\name{f.val.calc.Delta.obs.permil}
\alias{f.val.calc.Delta.obs.permil}
\title{Delta per mil, observed discrimination}

\usage{f.val.calc.Delta.obs.permil(reference.delta.e, chamber.delta.o, 
    xi)}
\arguments{
  \item{reference.delta.e}{}
  \item{chamber.delta.o}{}
  \item{xi}{}
}
\details{Multiply 1000 by \code{\link{f.val.calc.Delta.obs}}}
\value{Delta.obs.permil}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





