\name{f.val.calc.delta.permil}
\alias{f.val.calc.delta.permil}
\title{delta permil (Isotopic composition (d) per mil)}

\usage{f.val.calc.delta.permil(obs.12C, obs.13C, R.std)}
\arguments{
  \item{obs.12C}{}
  \item{obs.13C}{}
  \item{R.std}{}
}
\details{Multiply 1000 by \code{\link{f.val.calc.delta.proportion}}}
\value{delta.permil}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





