\name{f.val.calc.gtc}
\alias{f.val.calc.gtc}
\title{gtc, total (stomatal and boundary layer) conductance for CO2}

\usage{f.val.calc.gtc(gbc, gsc)}
\arguments{
  \item{gbc}{}
  \item{gsc}{}
}
\details{\deqn{gtc = 1 / ((1 / gbc) + (1 / gsc))}}
\value{gtc}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





