\name{f.val.calc.isotope.ratio}
\alias{f.val.calc.isotope.ratio}
\title{Isotope ratio}

\usage{f.val.calc.isotope.ratio(obs.12C, obs.13C)}
\arguments{
  \item{obs.12C}{}
  \item{obs.13C}{}
}
\details{\deqn{isotope.ratio = obs.13C / obs.12C}}
\value{isotope.ratio}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





