\name{f.val.calc.pc.using.complex.Delta.no.decarboxylation}
\alias{f.val.calc.pc.using.complex.Delta.no.decarboxylation}
\title{pc using complex D, no decarboxylation  [ab(pa-ps)+a(ps-pi)+pi(bs+al)-Dpa]/(bs+al-b)  is this different from two up? They give different values but both may not be derived properly}

\usage{f.val.calc.pc.using.complex.Delta.no.decarboxylation(Delta, pa, 
    ps, pi, a, a.b, al, b, bs)}
\arguments{
  \item{Delta}{}
  \item{pa}{}
  \item{ps}{}
  \item{pi}{}
  \item{a}{}
  \item{a.b}{}
  \item{al}{}
  \item{b}{}
  \item{bs}{}
}
\details{\deqn{pc = (a.b * (pa - ps) + a * (ps - pi) + pi * (bs + al) - Delta * pa) / (bs + al - b)}}
\value{pc}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





