\name{f.val.calc.pc.using.gm}
\alias{f.val.calc.pc.using.gm}
\title{pc using gm,  total partial pressure of CO2 at the site of carboxylation, Press is the atmospheric pressure value from the LI6400}

\usage{f.val.calc.pc.using.gm(pi, A, gm)}
\arguments{
  \item{pi}{}
  \item{A}{}
  \item{gm}{}
}
\details{\deqn{pc = pi - (A / gm)}}
\value{pc}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





