\name{f.val.calc.pc.using.simple.Delta.for.modeling}
\alias{f.val.calc.pc.using.simple.Delta.for.modeling}
\title{pc using simple D for modeling}

\usage{f.val.calc.pc.using.simple.Delta.for.modeling(Delta, pa, a, b.modeling)}
\arguments{
  \item{Delta}{}
  \item{pa}{}
  \item{a}{}
  \item{b.modeling}{}
}
\details{\deqn{pc = (pa * (Delta - a)) / (b.modeling - a)}}
\value{pc}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





