\name{f.val.calc.pp}
\alias{f.val.calc.pp}
\title{pa, partial pressure of CO2 above the leaf, Press is the atmospheric pressure value from the LI6400}

\usage{f.val.calc.pp(Co, Press)}
\arguments{
  \item{Co}{}
  \item{Press}{}
}
\details{\deqn{pa = (Co / 10^6) * (Press * 1000)}}
\value{pa}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





