\name{f.val.calc.TDL.A.photosynthesis}
\alias{f.val.calc.TDL.A.photosynthesis}
\title{A, TDL photosynthesis}

\usage{f.val.calc.TDL.A.photosynthesis(flow, Ce, Co, leaf.area)}
\arguments{
  \item{flow}{}
  \item{Ce}{}
  \item{Co}{}
  \item{leaf.area}{}
}
\details{\deqn{A = (flow / 10^6) * (Ce - Co) / (leaf.area / 10^4)}}
\value{A}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





