\name{f.val.calc.total.mol.fraction.CO2}
\alias{f.val.calc.total.mol.fraction.CO2}
\title{Total mol fraction CO2}

\usage{f.val.calc.total.mol.fraction.CO2(obs.12C, obs.13C, fo.13C)}
\arguments{
  \item{obs.12C}{}
  \item{obs.13C}{}
  \item{fo.13C}{}
}
\details{\deqn{total.mol.fraction.CO2 = (obs.12C + obs.13C) / (1 - fo.13C)}}
\value{total.mol.fraction.CO2}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





