\name{f.val.calc.xi}
\alias{f.val.calc.xi}
\title{xi}

\usage{f.val.calc.xi(Ce, Co)}
\arguments{
  \item{Ce}{}
  \item{Co}{}
}
\details{\deqn{xi = Ce / (Ce - Co)}}
\value{xi}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





