% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bs_iter_TDL_Licor.R
\name{f_bs_iter_TDL_Licor}
\alias{f_bs_iter_TDL_Licor}
\title{Title}
\usage{
f_bs_iter_TDL_Licor(val.obs.TDL, val.sum.TDL, val.obs.Licor, val.sum.Licor,
  var.interp)
}
\arguments{
\item{var.interp}{}
}
\description{
Title
}
