% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_np_bs_mean.R
\name{f_np_bs_mean}
\alias{f_np_bs_mean}
\title{calculate NP BS mean}
\usage{
f_np_bs_mean(x, first_ind, last_ind, n_sam, n)
}
\arguments{
\item{x}{xxxPARAMxxx}

\item{first_ind}{xxxPARAMxxx}

\item{last_ind}{xxxPARAMxxx}

\item{n_sam}{xxxPARAMxxx}

\item{n}{xxxPARAMxxx}
}
\value{
x_mean xxxRETURNxxx
}
\description{
Resample from data with replacement, calculating mean of n nonparametric bootstrap resamples.
}
