% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_par_bs_mean.R
\name{f_par_bs_mean}
\alias{f_par_bs_mean}
\title{calculate Par BS mean}
\usage{
f_par_bs_mean(x, x_var, n_sam, n)
}
\arguments{
\item{x}{xxxPARAMxxx}

\item{x_var}{xxxPARAMxxx}

\item{n_sam}{xxxPARAMxxx}

\item{n}{xxxPARAMxxx}
}
\value{
x_mean xxxRETURNxxx
}
\description{
Sample from normal distribution with mean and se from the data, calculating mean of n parametric bootstrap resamples.
}
