% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_bs_CI_endpoints.R
\name{f_val_bs_CI_endpoints}
\alias{f_val_bs_CI_endpoints}
\title{create CI endpoints}
\usage{
f_val_bs_CI_endpoints(x, n, ind_CI)
}
\arguments{
\item{ind_CI}{}
}
\value{
x_CI
}
\description{
Sorts BS samples, takes quantiles for equal-tailed sig_CI Level CIs.
}
