% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_corrected.R
\name{f_val_calc_corrected}
\alias{f_val_calc_corrected}
\title{Title}
\usage{
f_val_calc_corrected(obs.12C, obs.13C, gain_12C, gain_13C, offset_12C,
  offset_13C)
}
\arguments{
\item{offset_13C}{}
}
\description{
Title
}
