% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_from_A_ratio.R
\name{f_val_calc_Delta_from_A_ratio}
\alias{f_val_calc_Delta_from_A_ratio}
\title{Title}
\usage{
f_val_calc_Delta_from_A_ratio(cham.12Co, cham.13Co, TDL.13A, TDL.12A)
}
\arguments{
\item{TDL.12A}{}
}
\description{
Title
}
