% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_from_ratios_in_out.R
\name{f_val_calc_Delta_from_ratios_in_out}
\alias{f_val_calc_Delta_from_ratios_in_out}
\title{Delta from ratios in and out?, (Re/Ro)-1, is this really the same?}
\usage{
f_val_calc_Delta_from_ratios_in_out(ref_12Ce, ref_13Ce, cham_12Co,
  cham_13Co)
}
\arguments{
\item{ref_12Ce}{xxxPARAMxxx}

\item{ref_13Ce}{xxxPARAMxxx}

\item{cham_12Co}{xxxPARAMxxx}

\item{cham_13Co}{xxxPARAMxxx}
}
\value{
Delta_from_ratios_in_out xxxRETURNxxx
}
\description{
First calculates R_ref and R_cham with \code{\link{f_val_calc_isotope_ratio}}
}
\details{
\deqn{Delta_from_ratios_in_out <- (R_ref / R_cham) - 1}
this quantity is not in per mil
}
