% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_obs_permil.R
\name{f_val_calc_Delta_obs_permil}
\alias{f_val_calc_Delta_obs_permil}
\title{Delta per mil, observed discrimination}
\usage{
f_val_calc_Delta_obs_permil(reference_delta_e, chamber_delta_o, xi)
}
\arguments{
\item{xi}{}
}
\value{
Delta_obs_permil
}
\description{
Multiply 1000 by \code{\link{f_val_calc_Delta_obs}}
}
