% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_delta_proportion.R
\name{f_val_calc_delta_proportion}
\alias{f_val_calc_delta_proportion}
\title{Title}
\usage{
f_val_calc_delta_proportion(obs.12C, obs.13C, R_std)
}
\arguments{
\item{R_std}{}
}
\description{
Title
}
