% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_gain.R
\name{f_val_calc_gain}
\alias{f_val_calc_gain}
\title{Title}
\usage{
f_val_calc_gain(true.value.hi.12C, true.value.hi.13C, true.value.lo.12C,
  true.value.lo.13C, cal.tank.hi.12C, cal.tank.hi.13C, cal.tank.lo.12C,
  cal.tank.lo.13C)
}
\arguments{
\item{cal.tank.lo.13C}{}
}
\description{
Title
}
