% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_gtc.R
\name{f_val_calc_gtc}
\alias{f_val_calc_gtc}
\title{gtc, total (stomatal and boundary layer) conductance for CO2}
\usage{
f_val_calc_gtc(gbc, gsc)
}
\arguments{
\item{gbc}{xxxPARAMxxx}

\item{gsc}{xxxPARAMxxx}
}
\value{
gtc xxxRETURNxxx
}
\description{
\deqn{gtc = 1 / ((1 / gbc) + (1 / gsc))}
}
