% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_isotope_ratio.R
\name{f_val_calc_isotope_ratio}
\alias{f_val_calc_isotope_ratio}
\title{Isotope ratio}
\usage{
f_val_calc_isotope_ratio(obs_12C, obs_13C)
}
\arguments{
\item{obs_13C}{}
}
\value{
isotope_ratio
}
\description{
\deqn{isotope_ratio = obs_13C / obs_12C}
}
