% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_offset.R
\name{f_val_calc_offset}
\alias{f_val_calc_offset}
\title{offset (12C Offset, 13C Offset) using hi tank}
\usage{
f_val_calc_offset(true_value_hi_12C, true_value_hi_13C, gain_12C, gain_13C,
  cal_tank_hi_12C, cal_tank_hi_13C)
}
\arguments{
\item{true_value_hi_12C}{xxxPARAMxxx}

\item{true_value_hi_13C}{xxxPARAMxxx}

\item{gain_12C}{xxxPARAMxxx}

\item{gain_13C}{xxxPARAMxxx}

\item{cal_tank_hi_12C}{xxxPARAMxxx}

\item{cal_tank_hi_13C}{xxxPARAMxxx}
}
\value{
offset for 12C and 13C
}
\description{
\deqn{offset = true_value_hi - (gain * cal.tank_hi)}
}
