% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_p.R
\name{f_val_calc_p}
\alias{f_val_calc_p}
\title{p}
\usage{
f_val_calc_p(Ce, Co)
}
\arguments{
\item{Ce}{xxxPARAMxxx}

\item{Co}{xxxPARAMxxx}
}
\value{
p xxxRETURNxxx
}
\description{
\deqn{p = (Co-Ce) / Co}
}
