% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_complex_Delta_full_model.R
\name{f_val_calc_pc_using_complex_Delta_full_model}
\alias{f_val_calc_pc_using_complex_Delta_full_model}
\title{Title}
\usage{
f_val_calc_pc_using_complex_Delta_full_model(Delta, pa, ps, pi, a, a_b, al,
  b, bs, e, Rd, k, f, Gamma_star)
}
\arguments{
\item{Gamma_star}{}
}
\description{
Title
}
