% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_gm.R
\name{f_val_calc_pc_using_gm}
\alias{f_val_calc_pc_using_gm}
\title{pc using gm,  total partial pressure of CO2 at the site of carboxylation, Press is the atmospheric pressure value from the LI6400}
\usage{
f_val_calc_pc_using_gm(pi, A, gm)
}
\arguments{
\item{pi}{xxxPARAMxxx}

\item{A}{xxxPARAMxxx}

\item{gm}{xxxPARAMxxx}
}
\value{
pc xxxRETURNxxx
}
\description{
\deqn{pc = pi - (A / gm)}
}
