% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_simple_Delta_for_gm.R
\name{f_val_calc_pc_using_simple_Delta_for_gm}
\alias{f_val_calc_pc_using_simple_Delta_for_gm}
\title{Title}
\usage{
f_val_calc_pc_using_simple_Delta_for_gm(Delta, pa, ps, a, b.gm, a.b)
}
\arguments{
\item{a.b}{}
}
\description{
Title
}
