% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_simple_Delta_for_modeling.R
\name{f_val_calc_pc_using_simple_Delta_for_modeling}
\alias{f_val_calc_pc_using_simple_Delta_for_modeling}
\title{pc using simple D for modeling}
\usage{
f_val_calc_pc_using_simple_Delta_for_modeling(Delta, pa, a, b_modeling)
}
\arguments{
\item{Delta}{xxxPARAMxxx}

\item{pa}{xxxPARAMxxx}

\item{a}{xxxPARAMxxx}

\item{b_modeling}{xxxPARAMxxx}
}
\value{
pc xxxRETURNxxx
}
\description{
\deqn{pc = (pa * (Delta - a)) / (b_modeling - a)}
}
