% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_TDL_A_photosynthesis.R
\name{f_val_calc_TDL_A_photosynthesis}
\alias{f_val_calc_TDL_A_photosynthesis}
\title{A, TDL photosynthesis}
\usage{
f_val_calc_TDL_A_photosynthesis(flow, Ce, Co, leaf_area)
}
\arguments{
\item{leaf_area}{}
}
\value{
A
}
\description{
\deqn{A = (flow / 10^6) * (Ce - Co) / (leaf_area / 10^4)}
}
