% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_xi.R
\name{f_val_calc_xi}
\alias{f_val_calc_xi}
\title{xi}
\usage{
f_val_calc_xi(Ce, Co)
}
\arguments{
\item{Ce}{xxxPARAMxxx}

\item{Co}{xxxPARAMxxx}
}
\value{
xi xxxRETURNxxx
}
\description{
\deqn{xi = Ce / (Ce - Co)}
}
