\name{get.data}
\alias{get.data}
\title{Get data from Excel spreadsheet}

\usage{get.data(filename)}
\arguments{
  \item{filename}{}
}
\details{Read Excel template with \code{gdata} \code{read.xls} using perl.}
\value{DATA}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





