% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from Excel spreadsheet}
\usage{
get_data(filename)
}
\arguments{
\item{filename}{xxxPARAMxxx}
}
\value{
DATA xxxRETURNxxx
}
\description{
Read Excel template with \code{gdata} \code{read.xls} using perl.
}
