% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Prints 'Hello, world!' with "x".}
\usage{
hello(x = NULL, sep = " ")
}
\arguments{
\item{x}{A string}

\item{sep}{A string separator passed to \code{paste()}}
}
\value{
NULL, side effect is printing string.
}
\description{
\code{hello} returns two strings pasted together
}
\examples{
hello()
hello("Additional string.")
}
