% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_TDL_tanks_ref.R
\name{interp_TDL_tanks_ref}
\alias{interp_TDL_tanks_ref}
\title{Title}
\usage{
interp_TDL_tanks_ref(TDL, TDL_cycle, plot_format_list, output_fn_prefix,
  sw)
}
\arguments{
\item{sw}{}
}
\description{
Title
}
