% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isogasex.R
\name{isogasex}
\alias{isogasex}
\title{Reads TDL and Licor files, aligns them, calculates quantities of interest with bootstrap intervals.}
\usage{
isogasex(input_fn = c("isogasex_template4.yaml",
  "isogasex_template4.xls")[1], path = getwd())
}
\arguments{
\item{input_fn}{The parameter template file name.  A modified version of isogasex_template4.yaml (text) or isogasex_template4.xls (Excel worksheet)}

\item{path}{Directory path to input data files and the location where the ./out folder will be saved with all results.}
}
\value{
NULL invisibly, all results are saved as files in the ./out folder
}
\description{
Edit \file{isogasex_template4.yaml} or \file{isogasex_template4.xls} and input parameters and TDL/Licor filenames.
Run \code{\link{isogasex}}.
See output in \file{./out} directory.
}
\details{
SECTION Preamble
Set version.
Display header.

Save original working directory, change to data dir.
(This is a little hokey.)  Change to data dir, then to out dir, then back to original at end.

Create output directory.

SECTION Read data.

Input Excel workbook. \code{\link{read_template_param}}

Assign variables. \code{\link{assign_param_variables}}

Create output directory (prefix) and update the process_info.txt.

Check that template version matches version of isogasex.

Set random seed for samples.

Read TDL and Licor data.  \code{\link{read_TDL}} and \code{\link{read_Licor}}

SECTION TDL Plot and interpolate.

BEGIN If using TDL:

_ Print lines with missing values in TDL file and fix them. \code{\link{fix_missing_TDL_values}}

_ Extract cycle timing, write to file. \code{\link{extract_cycle_timing}}

_ Determine the last index for admissible values to calculate means. \code{\link{last_TDL_index_for_means}}

_ Plot each cycle of TDL measurements (visual diagnostics). \code{\link{plot_data_cycles}}

_ Calculate the mean and variance for the TDL data based on last measurements. \code{\link{calc_mean_TDL}}

_ Interpolate TDL tank and reference values. \code{\link{interp_TDL_tanks_ref}}

_ Summary values for TDL interpolated tanks and reference. \code{\link{calc_mean_TDL_interp_tanks_ref}}

_ Calculate mean Variance for interp values for Par BS of tanks and reference.

END

SECTION Align TDL and Licor

Determine overlapping time window of TDL and Licor measurements. \code{\link{align_TDL_Licor_times}}

Reduce data to overlapping time window and interp Licor values to TDL timepoints. \code{\link{time_window_TDL_Licor_interp}}

Plot each cycle of TDL measurements (visual diagnostics). \code{\link{plot_data_cycles}}

Calculate the mean and variance for the Licor data based on last measurements. \code{\link{calc_mean_Licor}}

SECTION Mean Calculations

Assign TDL and Licor values to val variable names, raw and summarized values. \code{\link{val_TDL_Licor_variables}}

summary values (mean). \code{\link{f_val_calc_all_driver}}

All time point values. \code{\link{f_val_calc_all_driver}}

SECTION Write out results

Summary files

TDL file. \code{\link{write_summary_TDL_file}}

Licor file. \code{\link{write_summary_Licor_file}}

Summary Calculation file. \code{\link{write_summary_Calc_file}}

All Calculation file. \code{\link{write_all_Calc_file}}

plot all the calculated values. \code{\link{plot_val_calc_sum}}

SECTION Bootstrap

BEGIN Bootstrap Calculations

_ Generate NP and Par BS samples for observed values.

. \code{\link{f_init_bs_matrix}}

. \code{\link{f_bs_iter_TDL_Licor}}

. \code{\link{f_bs_save_TDL_Licor}}

. \code{\link{f_val_calc_all_driver}}

. \code{\link{f_val_bs_matrix}}

_ calculate central 95\% intervals.

_ create CI for each calculated value. \code{\link{f_val_bs_CI}}

_ plot all variables with bs values, mean value, and CI intervals. \code{\link{f_plot_CI_individuals_driver}}

_ write out central 95\% interval, SD,

_ Summary files. \code{\link{write_CI_TDL_file}}

_ TDL file

_ Licor file. \code{\link{write_CI_Licor_file}}

_ Calculation file. \code{\link{write_CI_Calc_file}}

END

Save workspace.

SECTION Wrap it up.
Move selected plot files into subdirectories. \code{\link{move_plot_files}}

Copy template file to out dir.

Change back to original dir.

Complete.
}
