% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_TDL_index_for_means.R
\name{last_TDL_index_for_means}
\alias{last_TDL_index_for_means}
\title{Determine the last index for admissible values to calculate means}
\usage{
last_TDL_index_for_means(TDL, TDL_cycle)
}
\arguments{
\item{TDL}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}
}
\value{
TDL xxxRETURNxxx
}
\description{
Rind indices just before where site change occurs.
}
\details{
Error checking which sites are defined in template/TDL file but not the other.

Determine sites in TDL file not in template file.

Remove indices in TDL file not defined in the template (do not calculate for these).

Determine sites in template file not in TDL file.

Make sure the last TDL_cycle$table[,"last_n_obs"] are the same site as temp_last_ind.

If chamber, then go back and keep calculating index blocks.

For chamber, skip the first TDL_cycle$first_n_last_skip_chamber number of blocks.
Note that these are in reverse order, so the last rows in TDL$last_list are the first cycles.
Check that the last TDL_cycle$first_n_last_skip_chamber number of blocks are all chambers
and if so, then remove these cycles.

Sort by index.
}
