\name{move.plot.files}
\alias{move.plot.files}
\title{move selected plot files to subdirectories}

\usage{move.plot.files(plot.format)}
\arguments{
  \item{plot.format}{}
}
\details{Create \file{BS} and \file{var} directories.

Move \file{BS_CHECK*} files into \file{BS} directory.

Move files that are not (TDL, Licor, offset) into \file{var} directory.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





