% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_plot_files.R
\name{move_plot_files}
\alias{move_plot_files}
\title{move selected plot files to subdirectories}
\usage{
move_plot_files(plot_format)
}
\arguments{
\item{plot_format}{}
}
\description{
Create \file{BS} and \file{var} directories.
}
\details{
Move \file{BS_CHECK*} files into \file{BS} directory.

Move files that are not (TDL, Licor, offset) into \file{var} directory.
}
