\name{my.reg.fit}
\alias{my.reg.fit}
\title{basic SLR beta fit}

\usage{my.reg.fit(y, x)}
\arguments{
  \item{y}{}
  \item{x}{}
}
\details{Calculate intercept and slope of a simple linear regression fit.}
\value{beta.hat}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





