\name{my.reg.pred}
\alias{my.reg.pred}
\title{basic SLR y pred}

\usage{my.reg.pred(beta.hat, x)}
\arguments{
  \item{beta.hat}{}
  \item{x}{}
}
\details{Calculate predictions given the intercept and slope of a simple linear regression fit.}
\value{y.hat}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





