\name{plot.A.photosynthesis}
\alias{plot.A.photosynthesis}
\title{Plot A photosynthesis}

\usage{\method{plot}{A.photosynthesis}(TDL.A.photosynthesis, TDL.12A.photosynthesis, 
    TDL.13A.photosynthesis, Licor.A.photosynthesis, Delta.from.ratios.in.out, 
    Delta.from.A.ratio, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{TDL.A.photosynthesis}{}
  \item{TDL.12A.photosynthesis}{}
  \item{TDL.13A.photosynthesis}{}
  \item{Licor.A.photosynthesis}{}
  \item{Delta.from.ratios.in.out}{}
  \item{Delta.from.A.ratio}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





