\name{plot.Ca.Cs}
\alias{plot.Ca.Cs}
\title{Plot Ca Cs, CO2 concentrations above the leaf and at the leaf surface}

\usage{\method{plot}{Ca.Cs}(chamber.TotalCa, chamber.12Ca, chamber.13Ca, chamber.TotalCs, 
    chamber.12Cs, chamber.13Cs, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.TotalCa}{}
  \item{chamber.12Ca}{}
  \item{chamber.13Ca}{}
  \item{chamber.TotalCs}{}
  \item{chamber.12Cs}{}
  \item{chamber.13Cs}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





