\name{plot.Cc.total}
\alias{plot.Cc.total}
\title{Plot Cc, ppm CO2 concentration at the site of carboxylation, generally meaning inside the chloroplast and ignoring PEPC in cytosol}

\usage{\method{plot}{Cc.total}(chamber.TotalCc, chamber.12Cc, chamber.13Cc, x.time, 
    plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.TotalCc}{}
  \item{chamber.12Cc}{}
  \item{chamber.13Cc}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





