\name{plot.Ci.pi}
\alias{plot.Ci.pi}
\title{Plot Ci, pi, CO2 concentration and partial pressure of CO2 in the substomatal cavities}

\usage{\method{plot}{Ci.pi}(chamber.TotalCi, chamber.12Ci, chamber.13Ci, chamber.Totalpi, 
    chamber.12pi, chamber.13pi, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.TotalCi}{}
  \item{chamber.12Ci}{}
  \item{chamber.13Ci}{}
  \item{chamber.Totalpi}{}
  \item{chamber.12pi}{}
  \item{chamber.13pi}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





