\name{plot.corrected.total.diff.xi}
\alias{plot.corrected.total.diff.xi}
\title{Plot corrected 12CO2 and 13CO2, total, and difference for reference and sample}

\usage{\method{plot}{corrected.total.diff.xi}(ref.12C, ref.13C, cham.12C, cham.13C, 
    ref.tot, cham.tot, diff.tot, xi, x.time, plot.format.list, 
    output.fn.prefix)}
\arguments{
  \item{ref.12C}{}
  \item{ref.13C}{}
  \item{cham.12C}{}
  \item{cham.13C}{}
  \item{ref.tot}{}
  \item{cham.tot}{}
  \item{diff.tot}{}
  \item{xi}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





