\name{plot.delta.Delta.p}
\alias{plot.delta.Delta.p}
\title{Plot delta, Delta, and p}

\usage{\method{plot}{delta.Delta.p}(reference.delta.e, chamber.delta.o, chamber.reference.delta.diff.CoCe, 
    Delta.obs, p, delta.13C.Assim, delta.13C.Resp, x.time, plot.format.list, 
    output.fn.prefix)}
\arguments{
  \item{reference.delta.e}{}
  \item{chamber.delta.o}{}
  \item{chamber.reference.delta.diff.CoCe}{}
  \item{Delta.obs}{}
  \item{p}{}
  \item{delta.13C.Assim}{}
  \item{delta.13C.Resp}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





