\name{plot.Delta.i}
\alias{plot.Delta.i}
\title{Plot Delta.i, predicted discrimination}

\usage{\method{plot}{Delta.i}(chamber.Delta.i.simple.for.gm, chamber.Delta.i.simple.for.modeling, 
    chamber.Delta.i.complex.for.gm, chamber.Delta.i.simple.for.gm_Delta.obs, 
    chamber.Delta.i.complex.for.gm_Delta.obs, x.time, plot.format.list, 
    output.fn.prefix)}
\arguments{
  \item{chamber.Delta.i.simple.for.gm}{}
  \item{chamber.Delta.i.simple.for.modeling}{}
  \item{chamber.Delta.i.complex.for.gm}{}
  \item{chamber.Delta.i.simple.for.gm_Delta.obs}{}
  \item{chamber.Delta.i.complex.for.gm_Delta.obs}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





