\name{plot.gain.offset}
\alias{plot.gain.offset}
\title{Plot active gains, offsets, and corrected magnitudes}

\usage{\method{plot}{gain.offset}(gain.12C, gain.13C, offset.12C, offset.13C, 
    x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{gain.12C}{}
  \item{gain.13C}{}
  \item{offset.12C}{}
  \item{offset.13C}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





