\name{plot.gbc.gsc.gtc}
\alias{plot.gbc.gsc.gtc}
\title{Plot gbc, gsc, gtc: boundary layer, stomatal, and total conductance for CO2}

\usage{\method{plot}{gbc.gsc.gtc}(chamber.Totalgbc, chamber.13gbc, chamber.Totalgsc, 
    chamber.13gsc, chamber.Totalgtc, chamber.13gtc, x.time, plot.format.list, 
    output.fn.prefix)}
\arguments{
  \item{chamber.Totalgbc}{}
  \item{chamber.13gbc}{}
  \item{chamber.Totalgsc}{}
  \item{chamber.13gsc}{}
  \item{chamber.Totalgtc}{}
  \item{chamber.13gtc}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





