\name{plot.gm}
\alias{plot.gm}
\title{Plot gm point simple and complex, internal leaf (mesophyll) conductance calculated for every D value ignoring and estimating decarboxylation effects}

\usage{\method{plot}{gm}(chamber.Totalgm.point.simple, chamber.12gm.point.simple, 
    chamber.13gm.point.simple, chamber.Totalgm.point.complex, 
    x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.Totalgm.point.simple}{}
  \item{chamber.12gm.point.simple}{}
  \item{chamber.13gm.point.simple}{}
  \item{chamber.Totalgm.point.complex}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





