\name{plot.Licor.flow.press}
\alias{plot.Licor.flow.press}
\title{Plot Licor uin, xin, La, Atm.press values 7/21/2012}

\usage{\method{plot}{Licor.flow.press}(Licor.flow.uin, Licor.H2OR.xin, Licor.La, 
    Licor.Atm.press, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{Licor.flow.uin}{}
  \item{Licor.H2OR.xin}{}
  \item{Licor.La}{}
  \item{Licor.Atm.press}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





