\name{plot.Licor.temp.light}
\alias{plot.Licor.temp.light}
\title{Plot Licor Temp Light values}

\usage{\method{plot}{Licor.temp.light}(VPD, E.transpiration, leaf.temp, air.temp, 
    light.in, light.out, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{VPD}{}
  \item{E.transpiration}{}
  \item{leaf.temp}{}
  \item{air.temp}{}
  \item{light.in}{}
  \item{light.out}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





