\name{plot.not.na}
\alias{plot.not.na}
\title{if no data to plot, plot a dummy box (for when certain Licor columns are not collected)}

\usage{\method{plot}{not.na}(x.time, y.var, pch = 20, type, cex = 0.1, xlab = "", 
    ylab, main)}
\arguments{
  \item{x.time}{}
  \item{y.var}{}
  \item{pch}{}
  \item{type}{}
  \item{cex}{}
  \item{xlab}{}
  \item{ylab}{}
  \item{main}{}
}
\details{If no data to plot, returns a "no data" (0,0) point as plot place holder.}
\value{plot.not.na.val}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





