\name{plot.pa.ps}
\alias{plot.pa.ps}
\title{Plot Pa Ps, partial pressure of CO2 above the leaf and at the leaf surface}

\usage{\method{plot}{pa.ps}(chamber.Totalpa, chamber.12pa, chamber.13pa, chamber.Totalps, 
    chamber.12ps, chamber.13ps, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.Totalpa}{}
  \item{chamber.12pa}{}
  \item{chamber.13pa}{}
  \item{chamber.Totalps}{}
  \item{chamber.12ps}{}
  \item{chamber.13ps}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





