\name{plot.pc.simple.complex}
\alias{plot.pc.simple.complex}
\title{Plot pc, simple and complex}

\usage{\method{plot}{pc.simple.complex}(chamber.Totalpc.using.simple.Delta.for.gm, 
    chamber.Totalpc.using.simple.Delta.for.modeling, chamber.Totalpc.using.complex.Delta.no.decarboxylation, 
    chamber.Totalpc.using.complex.Delta.full.model, x.time, plot.format.list, 
    output.fn.prefix)}
\arguments{
  \item{chamber.Totalpc.using.simple.Delta.for.gm}{}
  \item{chamber.Totalpc.using.simple.Delta.for.modeling}{}
  \item{chamber.Totalpc.using.complex.Delta.no.decarboxylation}{}
  \item{chamber.Totalpc.using.complex.Delta.full.model}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





