\name{plot.pc.total}
\alias{plot.pc.total}
\title{Plot pc, total partial pressure of CO2 at the site of carboxylation}

\usage{\method{plot}{pc.total}(chamber.Totalpc.using.gm, chamber.12pc.using.gm, 
    chamber.13pc.using.gm, x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.Totalpc.using.gm}{}
  \item{chamber.12pc.using.gm}{}
  \item{chamber.13pc.using.gm}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





