\name{plot.pi_pa}
\alias{plot.pi_pa}
\title{Plot Totalpi_pa, ratio of substomatal CO2 partial pressure to CO2 partial pressure above leaf = Ci/Ca ratio of mol fractions}

\usage{\method{plot}{pi_pa}(chamber.Totalpi, chamber.Totalpa, chamber.Totalpi_pa, 
    x.time, plot.format.list, output.fn.prefix)}
\arguments{
  \item{chamber.Totalpi}{}
  \item{chamber.Totalpa}{}
  \item{chamber.Totalpi_pa}{}
  \item{x.time}{}
  \item{plot.format.list}{}
  \item{output.fn.prefix}{}
}
\details{Plots the input variables.}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





