% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_A_photosynthesis.R
\name{plot_A_photosynthesis}
\alias{plot_A_photosynthesis}
\title{Plot A photosynthesis}
\usage{
plot_A_photosynthesis(TDL_A_photosynthesis, TDL_12A_photosynthesis,
  TDL_13A_photosynthesis, Licor_A_photosynthesis, Delta_from_ratios_in_out,
  Delta_from_A_ratio, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{TDL_A_photosynthesis}{xxxPARAMxxx}

\item{TDL_12A_photosynthesis}{xxxPARAMxxx}

\item{TDL_13A_photosynthesis}{xxxPARAMxxx}

\item{Licor_A_photosynthesis}{xxxPARAMxxx}

\item{Delta_from_ratios_in_out}{xxxPARAMxxx}

\item{Delta_from_A_ratio}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
