% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Cc_total.R
\name{plot_Cc_total}
\alias{plot_Cc_total}
\title{Title}
\usage{
plot_Cc_total(chamber.TotalCc, chamber.12Cc, chamber.13Cc, x.time,
  plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
