% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ci_pi.R
\name{plot_Ci_pi}
\alias{plot_Ci_pi}
\title{Plot Ci, pi, CO2 concentration and partial pressure of CO2 in the substomatal cavities}
\usage{
plot_Ci_pi(chamber_TotalCi, chamber_12Ci, chamber_13Ci, chamber_Totalpi,
  chamber_12pi, chamber_13pi, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{chamber_TotalCi}{xxxPARAMxxx}

\item{chamber_12Ci}{xxxPARAMxxx}

\item{chamber_13Ci}{xxxPARAMxxx}

\item{chamber_Totalpi}{xxxPARAMxxx}

\item{chamber_12pi}{xxxPARAMxxx}

\item{chamber_13pi}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
