% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrected_total_diff_xi.R
\name{plot_corrected_total_diff_xi}
\alias{plot_corrected_total_diff_xi}
\title{Plot corrected 12CO2 and 13CO2, total, and difference for reference and sample}
\usage{
plot_corrected_total_diff_xi(ref_12C, ref_13C, cham_12C, cham_13C, ref_tot,
  cham_tot, diff_tot, xi, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{ref_12C}{xxxPARAMxxx}

\item{ref_13C}{xxxPARAMxxx}

\item{cham_12C}{xxxPARAMxxx}

\item{cham_13C}{xxxPARAMxxx}

\item{ref_tot}{xxxPARAMxxx}

\item{cham_tot}{xxxPARAMxxx}

\item{diff_tot}{xxxPARAMxxx}

\item{xi}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
