% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Delta_i.R
\name{plot_Delta_i}
\alias{plot_Delta_i}
\title{Plot Delta_i, predicted discrimination}
\usage{
plot_Delta_i(chamber_Delta_i_simple_for_gm,
  chamber_Delta_i_simple_for_modeling, chamber_Delta_i_complex_for_gm,
  chamber_Delta_i_simple_for_gm_Delta_obs,
  chamber_Delta_i_complex_for_gm_Delta_obs, x_time, plot_format_list,
  output_fn_prefix)
}
\arguments{
\item{chamber_Delta_i_simple_for_gm}{xxxPARAMxxx}

\item{chamber_Delta_i_simple_for_modeling}{xxxPARAMxxx}

\item{chamber_Delta_i_complex_for_gm}{xxxPARAMxxx}

\item{chamber_Delta_i_simple_for_gm_Delta_obs}{xxxPARAMxxx}

\item{chamber_Delta_i_complex_for_gm_Delta_obs}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
