% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gain_offset.R
\name{plot_gain_offset}
\alias{plot_gain_offset}
\title{Title}
\usage{
plot_gain_offset(gain.12C, gain.13C, offset.12C, offset.13C, x.time,
  plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
