% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gbc_gsc_gtc.R
\name{plot_gbc_gsc_gtc}
\alias{plot_gbc_gsc_gtc}
\title{Plot gbc, gsc, gtc: boundary layer, stomatal, and total conductance for CO2}
\usage{
plot_gbc_gsc_gtc(chamber_Totalgbc, chamber_13gbc, chamber_Totalgsc,
  chamber_13gsc, chamber_Totalgtc, chamber_13gtc, x_time, plot_format_list,
  output_fn_prefix)
}
\arguments{
\item{chamber_Totalgbc}{xxxPARAMxxx}

\item{chamber_13gbc}{xxxPARAMxxx}

\item{chamber_Totalgsc}{xxxPARAMxxx}

\item{chamber_13gsc}{xxxPARAMxxx}

\item{chamber_Totalgtc}{xxxPARAMxxx}

\item{chamber_13gtc}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
