% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Licor_temp_light.R
\name{plot_Licor_temp_light}
\alias{plot_Licor_temp_light}
\title{Title}
\usage{
plot_Licor_temp_light(VPD, E.transpiration, leaf.temp, air.temp, light.in,
  light.out, x.time, plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
