% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_not_na.R
\name{plot_not_na}
\alias{plot_not_na}
\title{if no data to plot, plot a dummy box (for when certain Licor columns are not collected)}
\usage{
plot_not_na(x_time, y_var, pch = 20, type, cex = 0.1, xlab = "",
  ylab, main)
}
\arguments{
\item{main}{}
}
\value{
plot_not_na_val
}
\description{
If no data to plot, returns a "no data" (0,0) point as plot place holder.
}
