% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pa_ps.R
\name{plot_pa_ps}
\alias{plot_pa_ps}
\title{Title}
\usage{
plot_pa_ps(chamber.Totalpa, chamber.12pa, chamber.13pa, chamber.Totalps,
  chamber.12ps, chamber.13ps, x.time, plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
