% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_simple_complex.R
\name{plot_pc_simple_complex}
\alias{plot_pc_simple_complex}
\title{Plot pc, simple and complex}
\usage{
plot_pc_simple_complex(chamber_Totalpc_using_simple_Delta_for_gm,
  chamber_Totalpc_using_simple_Delta_for_modeling,
  chamber_Totalpc_using_complex_Delta_no_decarboxylation,
  chamber_Totalpc_using_complex_Delta_full_model, x_time, plot_format_list,
  output_fn_prefix)
}
\arguments{
\item{chamber_Totalpc_using_simple_Delta_for_gm}{xxxPARAMxxx}

\item{chamber_Totalpc_using_simple_Delta_for_modeling}{xxxPARAMxxx}

\item{chamber_Totalpc_using_complex_Delta_no_decarboxylation}{xxxPARAMxxx}

\item{chamber_Totalpc_using_complex_Delta_full_model}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
