\name{progress.time}
\alias{progress.time}
\title{Calc amount of time since program started}

\usage{progress.time(time.start)}
\arguments{
  \item{time.start}{}
}
\details{Calculate time elapsed.}
\value{floor(time.sofar)}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





