% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_settings.R
\name{read_input_settings}
\alias{read_input_settings}
\title{Read input files for either training the classifier or for advert classification}
\usage{
read_input_settings(fn_scorecard = NULL, sw_read_write = c("read",
  "write"), out_dat = NULL, sw_log_file = NULL, out_log = NULL)
}
\arguments{
\item{fn_scorecard}{a list of csv filenames to read and join together, expected to have same column names.}

\item{sw_read_write}{"read" for infiles to train or classify, "write" for classification results (includes a column of classifications)}

\item{out_dat}{xxxPARAMxxx}

\item{sw_log_file}{see AT_ScoreCard_SigTest()}

\item{out_log}{defined inside AT_ScoreCard_SigTest()}
}
\value{
dat a single dataframe
}
\description{
Read input files for either training the classifier or for advert classification
}
