% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template_param.R
\name{read_template_param}
\alias{read_template_param}
\title{Read template parameter file}
\usage{
read_template_param(param_fn = NULL)
}
\arguments{
\item{param_fn}{A yaml or Excel parameter file}
}
\value{
PARAM_RAW a list of parameter values
}
\description{
Read template parameter file
}
