% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template_param_table_to_raw.R
\name{read_template_param_table_to_raw}
\alias{read_template_param_table_to_raw}
\title{Assign variables from xls template input file to yaml variable names}
\usage{
read_template_param_table_to_raw(D)
}
\arguments{
\item{D}{Parameter input table}
}
\value{
PARAM_RAW list of all inputs from Excel template
}
\description{
Assign variables from xls template input file to yaml variable names
}
