\name{s.plot.settings.begin.end}
\alias{s.plot.settings.begin.end}
\title{Plot settings for all plots}

\usage{s.plot.settings.begin.end(filename.prefix = "Prefix", plot.filename = "default_filename", 
    plot.mode = "settings", plot.format = 1)}
\arguments{
  \item{filename.prefix}{}
  \item{plot.filename}{}
  \item{plot.mode}{}
  \item{plot.format}{}
}
\details{Settings:

filenames

devices settings

par settings

Begin:

to interactive display

png is default output since quick display on web and used to convert to other raster formats

ps

pdf

bmp

jpeg

tiff

par settings

End:}
\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





