% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_plot_settings_begin_end.R
\name{s_plot_settings_begin_end}
\alias{s_plot_settings_begin_end}
\title{Title}
\usage{
s_plot_settings_begin_end(filename_prefix = "Prefix",
  plot_filename = "default_filename", plot_mode = "settings",
  plot_format = 1)
}
\arguments{
\item{plot_format}{}
}
\description{
Title
}
