\name{set.TDL.cycle}
\alias{set.TDL.cycle}
\title{TDL.cycle variables}

\usage{set.TDL.cycle(TDL.cycle)}
\arguments{
  \item{TDL.cycle}{}
}
\details{\code{last.n.obs.*} is the number of observations to take to be \code{seconds.*} long.}
\value{TDL.cycle}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





