% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_TDL_cycle.R
\name{set_TDL_cycle}
\alias{set_TDL_cycle}
\title{TDL_cycle variables}
\usage{
set_TDL_cycle(TDL_cycle)
}
\arguments{
\item{TDL_cycle}{xxxPARAMxxx}
}
\value{
TDL_cycle xxxRETURNxxx
}
\description{
\code{last_n_obs.*} is the number of observations to take to be \code{seconds.*} long.
}
