\name{tdllicor-package}
\alias{tdllicor-package}
\docType{package}
\title{TDL/Licor processing}
\description{Reads TDL and Licor files, aligns them, calculates quantities of interest with bootstrap intervals, and creates plots and tables.
Edit \file{tdllicor_templateX.xls} and input parameters and TDL/Licor filenames. Run \code{\link{tdllicor}}. See output in \file{./out} directory.}
\details{
\tabular{ll}{Package: \tab tdllicor\cr
Version: \tab 0.1-21\cr
Date: \tab 2012-10-17\cr
Type: \tab Package\cr
Title: \tab TDL/Licor processing\cr
Author: \tab Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>\cr
Maintainer: \tab Erik B. Erhardt <erike@stat.unm.edu>\cr
Depends: \tab R (>= 2.15.0),
gdata, zoo\cr
Suggests: \tab \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr}
}
\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}

\keyword{ package }


