\name{write.all.Calc.file}
\alias{write.all.Calc.file}
\title{write output section}
\description{Write all observed and calculated values for all time points.}
\usage{write.all.Calc.file(val, TDL.cycle, output.all.Calc.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.all.Calc.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





