\name{write.CI.Calc.file}
\alias{write.CI.Calc.file}
\title{write output section}
\description{Write all observed and calculated values with BS CIs for all time points.}
\usage{write.CI.Calc.file(val, TDL.cycle, output.CI.Calc.fn, output.CI.Calc.last.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.CI.Calc.fn}{}
  \item{output.CI.Calc.last.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





