\name{write.CI.Licor.file}
\alias{write.CI.Licor.file}
\title{write output section: Licor file}
\description{Write observed Licor values with BS CIs for all time points.}
\usage{write.CI.Licor.file(val, TDL.cycle, output.CI.Licor.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.CI.Licor.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





