\name{write.CI.TDL.file}
\alias{write.CI.TDL.file}
\title{write output section: TDL file}
\description{Write observed TDL values with BS CIs for all time points.}
\usage{write.CI.TDL.file(val, TDL.cycle, output.CI.TDL.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.CI.TDL.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





