\name{write.progress}
\alias{write.progress}
\title{write progress to the screen with time elapsed}

\usage{write.progress(text.to.cat, time.start, type.print = "cat")}
\arguments{
  \item{text.to.cat}{}
  \item{time.start}{}
  \item{type.print}{}
}

\value{NULL}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





