\name{write.summary.Calc.file}
\alias{write.summary.Calc.file}
\title{write output section}
\description{Write summary values for all time points.}
\usage{write.summary.Calc.file(val, TDL.cycle, output.summary.Calc.fn, 
    output.summary.Calc.last.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.summary.Calc.fn}{}
  \item{output.summary.Calc.last.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





