\name{write.summary.Licor.file}
\alias{write.summary.Licor.file}
\title{write output section: Licor file}
\description{Write Licor summary values for all time points.}
\usage{write.summary.Licor.file(val, TDL.cycle, output.summary.Licor.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.summary.Licor.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





