\name{write.summary.TDL.file}
\alias{write.summary.TDL.file}
\title{write output section: TDL file}
\description{Write TDL summary values for all time points.}
\usage{write.summary.TDL.file(val, TDL.cycle, output.summary.TDL.fn)}
\arguments{
  \item{val}{}
  \item{TDL.cycle}{}
  \item{output.summary.TDL.fn}{}
}

\value{val$write}

\author{Erik B. Erhardt <erike@stat.unm.edu> and David T. Hanson <dthanson@unm.edu>}





