% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_CI_Licor_file.R
\name{write_CI_Licor_file}
\alias{write_CI_Licor_file}
\title{write output section: Licor file}
\usage{
write_CI_Licor_file(val, TDL_cycle, output_CI_Licor_fn)
}
\arguments{
\item{val}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}

\item{output_CI_Licor_fn}{xxxPARAMxxx}
}
\value{
val$write xxxRETURNxxx
}
\description{
Write observed Licor values with BS CIs for all time points.
}
