% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_out.R
\name{write_out}
\alias{write_out}
\title{write progress to the screen}
\usage{
write_out(text_to_cat)
}
\arguments{
\item{text_to_cat}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
write progress to the screen
}
