% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_progress.R
\name{write_progress}
\alias{write_progress}
\title{write progress to the screen with time elapsed}
\usage{
write_progress(text_to_cat, time_start, type_print = c("cat", "matrix"))
}
\arguments{
\item{text_to_cat}{xxxPARAMxxx}

\item{time_start}{xxxPARAMxxx}

\item{type_print}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
write progress to the screen with time elapsed
}
