% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_Licor.R
\name{calc_mean_Licor}
\alias{calc_mean_Licor}
\title{Calculate the mean for the Licor data based on last measurements}
\usage{
calc_mean_Licor(TDL, Licor, TDL_cycle, sw)
}
\arguments{
\item{TDL}{xxxPARAMxxx}

\item{Licor}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}

\item{sw}{xxxPARAMxxx}
}
\value{
Licor xxxRETURNxxx
}
\description{
create Licor$summary for numerical summaries
}
\details{
Calculate mean for each site for retained observations
}
