% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cycle_timing.R
\name{extract_cycle_timing}
\alias{extract_cycle_timing}
\title{Cycle timing}
\usage{
extract_cycle_timing(TDL, TDL_cycle, TDL_site_timing_filename)
}
\arguments{
\item{TDL}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}

\item{TDL_site_timing_filename}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Write cycle timing to the timing file.
}
