% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cycle_timing.R
\name{extract_cycle_timing}
\alias{extract_cycle_timing}
\title{Title}
\usage{
extract_cycle_timing(TDL, TDL.cycle, TDL.site.timing.filename)
}
\arguments{
\item{TDL.site.timing.filename}{}
}
\description{
Title
}
