% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bs_iter_TDL_Licor.R
\name{f_bs_iter_TDL_Licor}
\alias{f_bs_iter_TDL_Licor}
\title{Title}
\usage{
f_bs_iter_TDL_Licor(val_obs_TDL, val_sum_TDL, val_obs_Licor, val_sum_Licor,
  var_interp)
}
\arguments{
\item{var_interp}{}
}
\description{
Title
}
