% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_np_bs_mean.R
\name{f_np_bs_mean}
\alias{f_np_bs_mean}
\title{calculate NP BS mean}
\usage{
f_np_bs_mean(x, first_ind, last_ind, n_sam, n)
}
\arguments{
\item{n}{}
}
\value{
x_mean
}
\description{
Resample from data with replacement, calculating mean of n nonparametric bootstrap resamples.
}
