% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_CI_individuals.R
\name{f_plot_CI_individuals}
\alias{f_plot_CI_individuals}
\title{plot all variables with bs values, mean value, and CI intervals}
\usage{
f_plot_CI_individuals(x_bs, x_sum, x_CI, x_title, x_time, x_n, R_bootstrap,
  plot_format_list, output_fn_prefix)
}
\arguments{
\item{x_bs}{xxxPARAMxxx}

\item{x_sum}{xxxPARAMxxx}

\item{x_CI}{xxxPARAMxxx}

\item{x_title}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{x_n}{xxxPARAMxxx}

\item{R_bootstrap}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
For every variable, plots the mean and bootstrap CI, after removing NAs with \code{\link{plot_not_na}}.
}
