% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_CI_individuals.R
\name{f_plot_CI_individuals}
\alias{f_plot_CI_individuals}
\title{Title}
\usage{
f_plot_CI_individuals(x_bs, x_sum, x_CI, x_title, x_time, x_n, R_bootstrap,
  plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Title
}
