% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_CI_individuals.R
\name{f_plot_CI_individuals}
\alias{f_plot_CI_individuals}
\title{Title}
\usage{
f_plot_CI_individuals(x.bs, x.sum, x.CI, x.title, x.time, x.n, R.bootstrap,
  plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
