% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_corrected.R
\name{f_val_calc_corrected}
\alias{f_val_calc_corrected}
\title{corrected (Corrected 12C, Corrected 13C)}
\usage{
f_val_calc_corrected(obs_12C, obs_13C, gain_12C, gain_13C, offset_12C,
  offset_13C)
}
\arguments{
\item{obs_12C}{xxxPARAMxxx}

\item{obs_13C}{xxxPARAMxxx}

\item{gain_12C}{xxxPARAMxxx}

\item{gain_13C}{xxxPARAMxxx}

\item{offset_12C}{xxxPARAMxxx}

\item{offset_13C}{xxxPARAMxxx}
}
\value{
corrected 12C and 13C
}
\description{
\deqn{corrected = (obs * gain) + offset}
}
