% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_delta_13C_Assim.R
\name{f_val_calc_delta_13C_Assim}
\alias{f_val_calc_delta_13C_Assim}
\title{Title}
\usage{
f_val_calc_delta_13C_Assim(chamber.delta.o, Delta.obs)
}
\arguments{
\item{Delta.obs}{}
}
\description{
Title
}
