% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_delta_13C_out_Resp.R
\name{f_val_calc_delta_13C_Resp}
\alias{f_val_calc_delta_13C_Resp}
\title{Title}
\usage{
f_val_calc_delta_13C_Resp(reference_delta_e, chamber_delta_o, p)
}
\arguments{
\item{p}{}
}
\description{
Title
}
