% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_from_A_ratio.R
\name{f_val_calc_Delta_from_A_ratio}
\alias{f_val_calc_Delta_from_A_ratio}
\title{D from A ratio, (Ro/(13A/12A))-1, should be the same as Dobs above}
\usage{
f_val_calc_Delta_from_A_ratio(cham_12Co, cham_13Co, TDL_13A, TDL_12A)
}
\arguments{
\item{cham_12Co}{xxxPARAMxxx}

\item{cham_13Co}{xxxPARAMxxx}

\item{TDL_13A}{xxxPARAMxxx}

\item{TDL_12A}{xxxPARAMxxx}
}
\value{
Delta_from_A_ratio xxxRETURNxxx
}
\description{
First calculates R_cham with \code{\link{f_val_calc_isotope_ratio}}
}
\details{
\deqn{Delta_from_A_ratio <- (R_cham / (TDL_13A / TDL_12A)) - 1}
this quantity is not in per mil
}
