% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_from_ratios_in_out.R
\name{f_val_calc_Delta_from_ratios_in_out}
\alias{f_val_calc_Delta_from_ratios_in_out}
\title{Title}
\usage{
f_val_calc_Delta_from_ratios_in_out(ref.12Ce, ref.13Ce, cham.12Co,
  cham.13Co)
}
\arguments{
\item{cham.13Co}{}
}
\description{
Title
}
