% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_i_complex_for_gm.R
\name{f_val_calc_Delta_i_complex_for_gm}
\alias{f_val_calc_Delta_i_complex_for_gm}
\title{Title}
\usage{
f_val_calc_Delta_i_complex_for_gm(a, b, a.b, pa, ps, pi, b.gm, f.photo,
  Gamma.star, e, Rd, k)
}
\arguments{
\item{k}{}
}
\description{
Title
}
