% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_i_simple_for_modeling.R
\name{f_val_calc_Delta_i_simple_for_modeling}
\alias{f_val_calc_Delta_i_simple_for_modeling}
\title{Title}
\usage{
f_val_calc_Delta_i_simple_for_modeling(a, b, pa, pi)
}
\arguments{
\item{pi}{}
}
\description{
Title
}
