% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_i_simple_for_modeling.R
\name{f_val_calc_Delta_i_simple_for_modeling}
\alias{f_val_calc_Delta_i_simple_for_modeling}
\title{Delta_i simple for modeling  a + (b-a) pi/pa predicted discrimination including boundary layer effects but using b adjustments to approximate effects of gm and decarboxylations}
\usage{
f_val_calc_Delta_i_simple_for_modeling(a, b, pa, pi)
}
\arguments{
\item{pi}{}
}
\value{
Delta_i_simple_for_modeling
}
\description{
\deqn{Delta_i_simple_for_modeling = a + (b - a) * pi / pa}
}
