% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_Delta_obs_permil.R
\name{f_val_calc_Delta_obs_permil}
\alias{f_val_calc_Delta_obs_permil}
\title{Title}
\usage{
f_val_calc_Delta_obs_permil(reference.delta.e, chamber.delta.o, xi)
}
\arguments{
\item{xi}{}
}
\description{
Title
}
