% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_delta_permil.R
\name{f_val_calc_delta_permil}
\alias{f_val_calc_delta_permil}
\title{delta permil (Isotopic composition (d) per mil)}
\usage{
f_val_calc_delta_permil(obs_12C, obs_13C, R_std)
}
\arguments{
\item{obs_12C}{xxxPARAMxxx}

\item{obs_13C}{xxxPARAMxxx}

\item{R_std}{xxxPARAMxxx}
}
\value{
delta_permil xxxRETURNxxx
}
\description{
Multiply 1000 by \code{\link{f_val_calc_delta_proportion}}
}
