% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_gm_point_simple.R
\name{f_val_calc_gm_point_simple}
\alias{f_val_calc_gm_point_simple}
\title{gm point simple, internal leaf (mesophyll) conductance calculated for every D value ignoring decarboxylation effects}
\usage{
f_val_calc_gm_point_simple(b, bs, al, A, pa, Di, D_obs)
}
\arguments{
\item{b}{xxxPARAMxxx}

\item{bs}{xxxPARAMxxx}

\item{al}{xxxPARAMxxx}

\item{A}{xxxPARAMxxx}

\item{pa}{xxxPARAMxxx}

\item{Di}{xxxPARAMxxx}

\item{D_obs}{xxxPARAMxxx}
}
\value{
gm_point_simple xxxRETURNxxx
}
\description{
\deqn{gm_point_simple = (( b - bs - al) * A / pa) / (Di - D_obs)}
}
