% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_gm_point_simple.R
\name{f_val_calc_gm_point_simple}
\alias{f_val_calc_gm_point_simple}
\title{Title}
\usage{
f_val_calc_gm_point_simple(b, bs, al, A, pa, Di, D_obs)
}
\arguments{
\item{D_obs}{}
}
\description{
Title
}
