% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_offset.R
\name{f_val_calc_offset}
\alias{f_val_calc_offset}
\title{Title}
\usage{
f_val_calc_offset(true.value.hi.12C, true.value.hi.13C, gain.12C, gain.13C,
  cal.tank.hi.12C, cal.tank.hi.13C)
}
\arguments{
\item{cal.tank.hi.13C}{}
}
\description{
Title
}
