% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_offset.R
\name{f_val_calc_offset}
\alias{f_val_calc_offset}
\title{Title}
\usage{
f_val_calc_offset(true_value_hi_12C, true_value_hi_13C, gain_12C, gain_13C,
  cal_tank_hi_12C, cal_tank_hi_13C)
}
\arguments{
\item{cal_tank_hi_13C}{}
}
\description{
Title
}
