% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_complex_Delta_full_model.R
\name{f_val_calc_pc_using_complex_Delta_full_model}
\alias{f_val_calc_pc_using_complex_Delta_full_model}
\title{pc using complex D, full model  [ab(pa-ps)+a(ps-pi)+pi(bs+al)-(eRd/k+fG*)-Dpa]/(bs+al-b)  includes boundary layer and decarboxylation effects}
\usage{
f_val_calc_pc_using_complex_Delta_full_model(Delta, pa, ps, pi, a, a_b, al,
  b, bs, e, Rd, k, f, Gamma_star)
}
\arguments{
\item{Gamma_star}{}
}
\value{
pc
}
\description{
\deqn{pc = (a_b * (pa - ps) + a * (ps - pi) + pi * (bs + al) - Delta * pa  - (e * Rd / k + f * Gamma_star)) / (bs + al - b)}
}
