% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/f_val_calc_pc_using_complex_Delta_no_decarboxylation.R
\name{f_val_calc_pc_using_complex_Delta_no_decarboxylation}
\alias{f_val_calc_pc_using_complex_Delta_no_decarboxylation}
\title{Title}
\usage{
f_val_calc_pc_using_complex_Delta_no_decarboxylation(Delta, pa, ps, pi, a,
  a_b, al, b, bs)
}
\arguments{
\item{bs}{}
}
\description{
Title
}
