% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_simple_Delta_for_gm.R
\name{f_val_calc_pc_using_simple_Delta_for_gm}
\alias{f_val_calc_pc_using_simple_Delta_for_gm}
\title{pc using simple D for gm, includes boundary layer}
\usage{
f_val_calc_pc_using_simple_Delta_for_gm(Delta, pa, ps, a, b_gm, a_b)
}
\arguments{
\item{Delta}{xxxPARAMxxx}

\item{pa}{xxxPARAMxxx}

\item{ps}{xxxPARAMxxx}

\item{a}{xxxPARAMxxx}

\item{b_gm}{xxxPARAMxxx}

\item{a_b}{xxxPARAMxxx}
}
\value{
pc xxxRETURNxxx
}
\description{
\deqn{pc = (Delta * pa - a_b * (pa - ps) - a * ps) / (b_gm - a)}
}
