% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_pc_using_simple_Delta_for_modeling.R
\name{f_val_calc_pc_using_simple_Delta_for_modeling}
\alias{f_val_calc_pc_using_simple_Delta_for_modeling}
\title{Title}
\usage{
f_val_calc_pc_using_simple_Delta_for_modeling(Delta, pa, a, b.modeling)
}
\arguments{
\item{b.modeling}{}
}
\description{
Title
}
