% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_TDL_A_photosynthesis.R
\name{f_val_calc_TDL_A_photosynthesis}
\alias{f_val_calc_TDL_A_photosynthesis}
\title{Title}
\usage{
f_val_calc_TDL_A_photosynthesis(flow, Ce, Co, leaf.area)
}
\arguments{
\item{leaf.area}{}
}
\description{
Title
}
