% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_val_calc_total_mol_fraction_CO2.R
\name{f_val_calc_total_mol_fraction_CO2}
\alias{f_val_calc_total_mol_fraction_CO2}
\title{Title}
\usage{
f_val_calc_total_mol_fraction_CO2(obs.12C, obs.13C, fo.13C)
}
\arguments{
\item{fo.13C}{}
}
\description{
Title
}
