% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isogasex_logo.R
\name{isogasex_logo}
\alias{isogasex_logo}
\title{isogasex logo, just for fun}
\usage{
isogasex_logo(sw_unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{sw_unicode}{}
}
\description{
isogasex logo, just for fun
}
