% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_reg_pred.R
\name{my_reg_pred}
\alias{my_reg_pred}
\title{basic SLR y pred}
\usage{
my_reg_pred(beta_hat, x)
}
\arguments{
\item{beta_hat}{xxxPARAMxxx}

\item{x}{xxxPARAMxxx}
}
\value{
y_hat xxxRETURNxxx
}
\description{
Calculate predictions given the intercept and slope of a simple linear regression fit.
}
