% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Cc_total.R
\name{plot_Cc_total}
\alias{plot_Cc_total}
\title{Title}
\usage{
plot_Cc_total(chamber_TotalCc, chamber_12Cc, chamber_13Cc, x_time,
  plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Title
}
