% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ci_pi.R
\name{plot_Ci_pi}
\alias{plot_Ci_pi}
\title{Title}
\usage{
plot_Ci_pi(chamber_TotalCi, chamber_12Ci, chamber_13Ci, chamber_Totalpi,
  chamber_12pi, chamber_13pi, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Title
}
