% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrected_total_diff_xi.R
\name{plot_corrected_total_diff_xi}
\alias{plot_corrected_total_diff_xi}
\title{Title}
\usage{
plot_corrected_total_diff_xi(ref.12C, ref.13C, cham.12C, cham.13C, ref_tot,
  cham_tot, diff_tot, xi, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Title
}
