% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrected_total_diff_xi.R
\name{plot_corrected_total_diff_xi}
\alias{plot_corrected_total_diff_xi}
\title{Title}
\usage{
plot_corrected_total_diff_xi(ref.12C, ref.13C, cham.12C, cham.13C, ref.tot,
  cham.tot, diff.tot, xi, x.time, plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
