% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Delta_i.R
\name{plot_Delta_i}
\alias{plot_Delta_i}
\title{Title}
\usage{
plot_Delta_i(chamber.Delta.i.simple.for.gm,
  chamber.Delta.i.simple.for.modeling, chamber.Delta.i.complex.for.gm,
  chamber.Delta.i.simple.for.gm_Delta.obs,
  chamber.Delta.i.complex.for.gm_Delta.obs, x.time, plot.format.list,
  output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
