% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gain_offset.R
\name{plot_gain_offset}
\alias{plot_gain_offset}
\title{Plot active gains, offsets, and corrected magnitudes}
\usage{
plot_gain_offset(gain_12C, gain_13C, offset_12C, offset_13C, x_time,
  plot_format_list, output_fn_prefix)
}
\arguments{
\item{gain_12C}{xxxPARAMxxx}

\item{gain_13C}{xxxPARAMxxx}

\item{offset_12C}{xxxPARAMxxx}

\item{offset_13C}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
