% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gain_offset.R
\name{plot_gain_offset}
\alias{plot_gain_offset}
\title{Title}
\usage{
plot_gain_offset(gain_12C, gain_13C, offset_12C, offset_13C, x_time,
  plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Title
}
