% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gbc_gsc_gtc.R
\name{plot_gbc_gsc_gtc}
\alias{plot_gbc_gsc_gtc}
\title{Title}
\usage{
plot_gbc_gsc_gtc(chamber.Totalgbc, chamber.13gbc, chamber.Totalgsc,
  chamber.13gsc, chamber.Totalgtc, chamber.13gtc, x.time, plot.format.list,
  output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
