% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gbc_gsc_gtc.R
\name{plot_gbc_gsc_gtc}
\alias{plot_gbc_gsc_gtc}
\title{Title}
\usage{
plot_gbc_gsc_gtc(chamber_Totalgbc, chamber_13gbc, chamber_Totalgsc,
  chamber_13gsc, chamber_Totalgtc, chamber_13gtc, x_time, plot_format_list,
  output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Title
}
