% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gm.R
\name{plot_gm}
\alias{plot_gm}
\title{Title}
\usage{
plot_gm(chamber.Totalgm.point.simple, chamber.12gm.point.simple,
  chamber.13gm.point.simple, chamber.Totalgm.point.complex, x.time,
  plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
