% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gm.R
\name{plot_gm}
\alias{plot_gm}
\title{Plot gm point simple and complex, internal leaf (mesophyll) conductance calculated for every D value ignoring and estimating decarboxylation effects}
\usage{
plot_gm(chamber_Totalgm_point_simple, chamber_12gm_point_simple,
  chamber_13gm_point_simple, chamber_Totalgm_point_complex, x_time,
  plot_format_list, output_fn_prefix)
}
\arguments{
\item{chamber_Totalgm_point_simple}{xxxPARAMxxx}

\item{chamber_12gm_point_simple}{xxxPARAMxxx}

\item{chamber_13gm_point_simple}{xxxPARAMxxx}

\item{chamber_Totalgm_point_complex}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
