% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Licor_temp_light.R
\name{plot_Licor_temp_light}
\alias{plot_Licor_temp_light}
\title{Plot Licor Temp Light values}
\usage{
plot_Licor_temp_light(VPD, E_transpiration, leaf_temp, air_temp, light_in,
  light_out, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{VPD}{xxxPARAMxxx}

\item{E_transpiration}{xxxPARAMxxx}

\item{leaf_temp}{xxxPARAMxxx}

\item{air_temp}{xxxPARAMxxx}

\item{light_in}{xxxPARAMxxx}

\item{light_out}{xxxPARAMxxx}

\item{x_time}{xxxPARAMxxx}

\item{plot_format_list}{xxxPARAMxxx}

\item{output_fn_prefix}{xxxPARAMxxx}
}
\value{
NULL xxxRETURNxxx
}
\description{
Plots the input variables.
}
