% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pa_ps.R
\name{plot_pa_ps}
\alias{plot_pa_ps}
\title{Plot Pa Ps, partial pressure of CO2 above the leaf and at the leaf surface}
\usage{
plot_pa_ps(chamber_Totalpa, chamber_12pa, chamber_13pa, chamber_Totalps,
  chamber_12ps, chamber_13ps, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Plots the input variables.
}
