% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_simple_complex.R
\name{plot_pc_simple_complex}
\alias{plot_pc_simple_complex}
\title{Title}
\usage{
plot_pc_simple_complex(chamber.Totalpc.using.simple.Delta.for.gm,
  chamber.Totalpc.using.simple.Delta.for.modeling,
  chamber.Totalpc.using.complex.Delta.no.decarboxylation,
  chamber.Totalpc.using.complex.Delta.full.model, x.time, plot.format.list,
  output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
