% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_total.R
\name{plot_pc_total}
\alias{plot_pc_total}
\title{Plot pc, total partial pressure of CO2 at the site of carboxylation}
\usage{
plot_pc_total(chamber_Totalpc_using_gm, chamber_12pc_using_gm,
  chamber_13pc_using_gm, x_time, plot_format_list, output_fn_prefix)
}
\arguments{
\item{output_fn_prefix}{}
}
\description{
Plots the input variables.
}
