% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_total.R
\name{plot_pc_total}
\alias{plot_pc_total}
\title{Title}
\usage{
plot_pc_total(chamber.Totalpc.using.gm, chamber.12pc.using.gm,
  chamber.13pc.using.gm, x.time, plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
