% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pi_pa.R
\name{plot_pi_pa}
\alias{plot_pi_pa}
\title{Title}
\usage{
plot_pi_pa(chamber.Totalpi, chamber.Totalpa, chamber.Totalpi_pa, x.time,
  plot.format.list, output.fn.prefix)
}
\arguments{
\item{output.fn.prefix}{}
}
\description{
Title
}
