% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_val_calc_sum.R
\name{plot_val_calc_sum}
\alias{plot_val_calc_sum}
\title{Title}
\usage{
plot_val_calc_sum(val.TDL, val.Licor, val.const, val.temp,
  plot.format.list, output.fn.prefix, sw)
}
\arguments{
\item{sw}{}
}
\description{
Title
}
