% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_plot_settings_begin_end.R
\name{s_plot_settings_begin_end}
\alias{s_plot_settings_begin_end}
\title{Title}
\usage{
s_plot_settings_begin_end(filename.prefix = "Prefix",
  plot.filename = "default_filename", plot.mode = "settings",
  plot.format = 1)
}
\arguments{
\item{plot.format}{}
}
\description{
Title
}
