% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_plot_settings_begin_end.R
\name{s_plot_settings_begin_end}
\alias{s_plot_settings_begin_end}
\title{Plot settings for all plots}
\usage{
s_plot_settings_begin_end(filename_prefix = "Prefix",
  plot_filename = "default_filename", plot_mode = "settings",
  plot_format = 1)
}
\arguments{
\item{plot_format}{}
}
\description{
Settings:
}
\details{
filenames

devices settings

par settings

Begin:

to interactive display

png is default output since quick display on web and used to convert to other raster formats

ps

pdf

bmp

jpeg

tiff

par settings

End:
}
