% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_window_TDL_Licor_interp.R
\name{time_window_TDL_Licor_interp}
\alias{time_window_TDL_Licor_interp}
\title{Title}
\usage{
time_window_TDL_Licor_interp(TDL, Licor, TDL.Licor.times, sw)
}
\arguments{
\item{sw}{}
}
\description{
Title
}
