% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_all_Calc_file.R
\name{write_all_Calc_file}
\alias{write_all_Calc_file}
\title{write output section}
\usage{
write_all_Calc_file(val, TDL_cycle, output_all_Calc_fn)
}
\arguments{
\item{output_all_Calc_fn}{}
}
\value{
val$write
}
\description{
Write all observed and calculated values for all time points.
}
