% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_CI_Calc_file.R
\name{write_CI_Calc_file}
\alias{write_CI_Calc_file}
\title{Title}
\usage{
write_CI_Calc_file(val, TDL.cycle, output.CI.Calc.fn,
  output.CI.Calc.last.fn)
}
\arguments{
\item{output.CI.Calc.last.fn}{}
}
\description{
Title
}
