% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_CI_Calc_file.R
\name{write_CI_Calc_file}
\alias{write_CI_Calc_file}
\title{write output section}
\usage{
write_CI_Calc_file(val, TDL_cycle, output_CI_Calc_fn,
  output_CI_Calc_last_fn)
}
\arguments{
\item{val}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}

\item{output_CI_Calc_fn}{xxxPARAMxxx}

\item{output_CI_Calc_last_fn}{xxxPARAMxxx}
}
\value{
val$write xxxRETURNxxx
}
\description{
Write all observed and calculated values with BS CIs for all time points.
}
