% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_CI_TDL_file.R
\name{write_CI_TDL_file}
\alias{write_CI_TDL_file}
\title{write output section: TDL file}
\usage{
write_CI_TDL_file(val, TDL_cycle, output_CI_TDL_fn)
}
\arguments{
\item{output_CI_TDL_fn}{}
}
\value{
val$write
}
\description{
Write observed TDL values with BS CIs for all time points.
}
