% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_Licor_file.R
\name{write_summary_Licor_file}
\alias{write_summary_Licor_file}
\title{write output section: Licor file}
\usage{
write_summary_Licor_file(val, TDL_cycle, output_summary_Licor_fn)
}
\arguments{
\item{val}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}

\item{output_summary_Licor_fn}{xxxPARAMxxx}
}
\value{
val$write xxxRETURNxxx
}
\description{
write output section: Licor file
}
