% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_TDL_file.R
\name{write_summary_TDL_file}
\alias{write_summary_TDL_file}
\title{Title}
\usage{
write_summary_TDL_file(val, TDL.cycle, output.summary.TDL.fn)
}
\arguments{
\item{output.summary.TDL.fn}{}
}
\description{
Title
}
