% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_TDL_file.R
\name{write_summary_TDL_file}
\alias{write_summary_TDL_file}
\title{write output section: TDL file}
\usage{
write_summary_TDL_file(val, TDL_cycle, output_summary_TDL_fn)
}
\arguments{
\item{val}{xxxPARAMxxx}

\item{TDL_cycle}{xxxPARAMxxx}

\item{output_summary_TDL_fn}{xxxPARAMxxx}
}
\value{
val$write xxxRETURNxxx
}
\description{
Write TDL summary values for all time points.
}
